New version of the NWChem parsers.
This is the first draft of the refactored NWCHem parsers. The parsing
now uses a modular approach so that the context that the templates are
oerating in is more obvious. The parsing now returns a nested tree of
calculations, each with their own initialization and finalization modules
so that it is simple to query (e.g.) the total_energy through a whole
calculation, so that a plot of all the energy points in a geometry
optimisatin becomes relatively straightforward.

Units are now added to all terms that require them in the compchem
dictionary. The compchem namespace is now used in preference to cc.

The Template unit tests have been updated, and grouped by those
that work and are tested, those that are used, but for various
reasons cannot be tested (usu. whitespace issues) and those that
are not used and untested and need to be re-integrated back into
the new strategy.

The convention test has also been updated.

The new templates currently work for SCF, DFT and MP2 single-point
calculations and geometry optimisations, and will extract the energies
and times for these calcualtions.

Other NWChem runs should be parsed, although only molecular structures
and basis_sets, together with the environment information will  be extracted.